Warning fixes

+++ libfreerdp/primitives/test/TestPrimitivesYUV.c
@@ -635,7 +635,7 @@ static BOOL check_for_mismatches(const BYTE* planeA, const BYTE* planeB, UINT32 
 		const BYTE a = planeA[x];
 		const BYTE b = planeB[x];
 
-		if (fabsf(a - b) > 2.0f)
+		if (fabsf((float)a - (float)b) > 2.0f)
 		{
 			rc = TRUE;
 			fprintf(stderr, "[%08x] %02x != %02x\n",
